requirements.txt
absl-py==1.3.0
attrs==22.1.0
contourpy==1.0.6
cycler==0.11.0
docutils==0.19
exceptiongroup==1.0.1
fonttools==4.38.0
iniconfig==1.1.1
kiwisolver==1.4.4
m2r==0.2.1
matplotlib==3.6.2
mistune==2.0.4
networkx==2.8.8
numpy==1.23.4
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.6
scipy==1.9.3
six==1.16.0
tomli==2.0.1
lifelines==0.27.4
py-spy==0.3.14
cloudpickle==2.2.0
ml-collections==0.1.1